use float32 types for all interfaces

+++ src/cpp/curve_network.cpp
@@ -35,8 +35,8 @@ void bind_curve_network(py::module& m) {
   bindStructure<ps::CurveNetwork>(m, "CurveNetwork")
 
     // basics
-    .def("update_node_positions", &ps::CurveNetwork::updateNodePositions<Eigen::MatrixXd>, "Update node positions")
-    .def("update_node_positions2D", &ps::CurveNetwork::updateNodePositions2D<Eigen::MatrixXd>, "Update node positions")
+    .def("update_node_positions", &ps::CurveNetwork::updateNodePositions<Eigen::MatrixXf>, "Update node positions")
+    .def("update_node_positions2D", &ps::CurveNetwork::updateNodePositions2D<Eigen::MatrixXf>, "Update node positions")
     .def("n_nodes", &ps::CurveNetwork::nNodes, "# nodes")
     .def("n_edges", &ps::CurveNetwork::nEdges, "# edges")
 
@@ -49,42 +49,42 @@ void bind_curve_network(py::module& m) {
     .def("get_material", &ps::CurveNetwork::getMaterial, "Get material")
     
     // quantities
-    .def("add_node_color_quantity", &ps::CurveNetwork::addNodeColorQuantity<Eigen::MatrixXd>, "Add a color function at nodes",
+    .def("add_node_color_quantity", &ps::CurveNetwork::addNodeColorQuantity<Eigen::MatrixXf>, "Add a color function at nodes",
         py::arg("name"), py::arg("values"), py::return_value_policy::reference)
-    .def("add_edge_color_quantity", &ps::CurveNetwork::addEdgeColorQuantity<Eigen::MatrixXd>, "Add a color function at edges",
+    .def("add_edge_color_quantity", &ps::CurveNetwork::addEdgeColorQuantity<Eigen::MatrixXf>, "Add a color function at edges",
         py::arg("name"), py::arg("values"), py::return_value_policy::reference)
-    .def("add_node_scalar_quantity", &ps::CurveNetwork::addNodeScalarQuantity<Eigen::VectorXd>, "Add a scalar function at nodes",
+    .def("add_node_scalar_quantity", &ps::CurveNetwork::addNodeScalarQuantity<Eigen::VectorXf>, "Add a scalar function at nodes",
         py::arg("name"), py::arg("values"), py::arg("data_type")=ps::DataType::STANDARD, py::return_value_policy::reference)
-    .def("add_edge_scalar_quantity", &ps::CurveNetwork::addEdgeScalarQuantity<Eigen::VectorXd>, "Add a scalar function at edge",
+    .def("add_edge_scalar_quantity", &ps::CurveNetwork::addEdgeScalarQuantity<Eigen::VectorXf>, "Add a scalar function at edge",
         py::arg("name"), py::arg("values"), py::arg("data_type")=ps::DataType::STANDARD, py::return_value_policy::reference)
-    .def("add_node_vector_quantity", &ps::CurveNetwork::addNodeVectorQuantity<Eigen::MatrixXd>, "Add a vector function at nodes",
+    .def("add_node_vector_quantity", &ps::CurveNetwork::addNodeVectorQuantity<Eigen::MatrixXf>, "Add a vector function at nodes",
         py::arg("name"), py::arg("values"), py::arg("vector_type")=ps::VectorType::STANDARD, py::return_value_policy::reference)
-    .def("add_node_vector_quantity2D", &ps::CurveNetwork::addNodeVectorQuantity2D<Eigen::MatrixXd>, "Add a vector function at nodes",
+    .def("add_node_vector_quantity2D", &ps::CurveNetwork::addNodeVectorQuantity2D<Eigen::MatrixXf>, "Add a vector function at nodes",
         py::arg("name"), py::arg("values"), py::arg("vector_type")=ps::VectorType::STANDARD, py::return_value_policy::reference)
-    .def("add_edge_vector_quantity", &ps::CurveNetwork::addEdgeVectorQuantity<Eigen::MatrixXd>, "Add a vector function at edges",
+    .def("add_edge_vector_quantity", &ps::CurveNetwork::addEdgeVectorQuantity<Eigen::MatrixXf>, "Add a vector function at edges",
         py::arg("name"), py::arg("values"), py::arg("vector_type")=ps::VectorType::STANDARD, py::return_value_policy::reference)
-    .def("add_edge_vector_quantity2D", &ps::CurveNetwork::addEdgeVectorQuantity2D<Eigen::MatrixXd>, "Add a vector function at edges",
+    .def("add_edge_vector_quantity2D", &ps::CurveNetwork::addEdgeVectorQuantity2D<Eigen::MatrixXf>, "Add a vector function at edges",
         py::arg("name"), py::arg("values"), py::arg("vector_type")=ps::VectorType::STANDARD, py::return_value_policy::reference);
 
   // Static adders and getters
-  m.def("register_curve_network", &ps::registerCurveNetwork<Eigen::MatrixXd, Eigen::MatrixXi>, 
+  m.def("register_curve_network", &ps::registerCurveNetwork<Eigen::MatrixXf, Eigen::MatrixXi>, 
       py::arg("name"), py::arg("nodes"), py::arg("edges"),
       "Register a curve network", py::return_value_policy::reference);
-  m.def("register_curve_network2D", &ps::registerCurveNetwork2D<Eigen::MatrixXd, Eigen::MatrixXi>, 
+  m.def("register_curve_network2D", &ps::registerCurveNetwork2D<Eigen::MatrixXf, Eigen::MatrixXi>, 
       py::arg("name"), py::arg("nodes"), py::arg("edges"),
       "Register a curve network", py::return_value_policy::reference);
   
-  m.def("register_curve_network_line", &ps::registerCurveNetworkLine<Eigen::MatrixXd>, 
+  m.def("register_curve_network_line", &ps::registerCurveNetworkLine<Eigen::MatrixXf>, 
       py::arg("name"), py::arg("nodes"), 
       "Register a curve network", py::return_value_policy::reference);
-  m.def("register_curve_network_line2D", &ps::registerCurveNetworkLine2D<Eigen::MatrixXd>, 
+  m.def("register_curve_network_line2D", &ps::registerCurveNetworkLine2D<Eigen::MatrixXf>, 
       py::arg("name"), py::arg("nodes"),
       "Register a curve network", py::return_value_policy::reference);
   
-  m.def("register_curve_network_loop", &ps::registerCurveNetworkLoop<Eigen::MatrixXd>, 
+  m.def("register_curve_network_loop", &ps::registerCurveNetworkLoop<Eigen::MatrixXf>, 
       py::arg("name"), py::arg("nodes"),
       "Register a curve network", py::return_value_policy::reference);
-  m.def("register_curve_network_loop2D", &ps::registerCurveNetworkLoop2D<Eigen::MatrixXd>, 
+  m.def("register_curve_network_loop2D", &ps::registerCurveNetworkLoop2D<Eigen::MatrixXf>, 
       py::arg("name"), py::arg("nodes"), 
       "Register a curve network", py::return_value_policy::reference);
